new: jwt config
error fixing

+++ ravine/src/main/java/liqour/ravine/entities/Customer.java
@@ -25,7 +25,7 @@ public class Customer {
     private String email;
     
     @Column(name="phone_number")
-    private long phone_number;
+    private int phone_number;
     
     @Column(name="image_url")
     private String image_url;
